Develop payment integratio

+++ fleet-management/src/main/java/com/fleet/management/dto/tarjetacombustible/TarjetaCombustibleResponse.java
@@ -5,6 +5,7 @@ import com.fleet.management.dto.empresa.EmpresaResponse;
 import com.fleet.management.dto.user.UserAuditResponse;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -16,7 +17,8 @@ public class TarjetaCombustibleResponse {
 
     private Long id;
     private String numero;
-    private Double saldo;
+    // FX-13: migrado de Double a BigDecimal para precision monetaria.
+    private BigDecimal saldo;
     private CurrencyResponse currency;
     private EmpresaResponse empresa;
     private Boolean activo;
